Refactor TickerPanel into dedicated component
